training/requirements.txt
h5py>=3.10
numpy>=1.24
Pillow>=10.0
